requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
decorator==4.4.2
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
dynaconf==3.1.2
hvac==0.10.8
idna==2.10
knox==0.1.11
loguru==0.5.3
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pytz==2021.1
requests==2.25.1
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
validators==0.18.2
